+++ lib/protoflow-blocks/src/lib.rs
@@ -65,8 +65,10 @@ pub fn build_stdio_system(
         "Hash" => Hash::build_system(config)?,
         // IoBlocks
         "Decode" => Decode::build_system(config)?,
+        "DecodeJSON" => DecodeJson::build_system(config)?,
         "Encode" => Encode::build_system(config)?,
         "EncodeHex" => EncodeHex::build_system(config)?,
+        "EncodeJSON" => EncodeJson::build_system(config)?,
         // MathBlocks
         // SysBlocks
         "ReadDir" => ReadDir::build_system(config)?,