+++ CHANGES.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.3] - 2024-11-15
+### Fixed
+- Feature flags
+
+## [0.4.2] - 2024-11-15
+### Added
+- `DecodeJson` block
+- `ReadFile` block
+- `WriteFile` block
+
+## [0.4.1] - 2024-10-09
+### Added
+- `EncodeJson` block
+- Serde support for core types
+### Fixed
+- CLI: Fix a minor bug
+
 ## [0.4.0] - 2024-10-01
 
 ## [0.3.1] - 2024-09-25
@@ -51,6 +68,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2024-08-20
 
+[0.4.3]: https://github.com/AsimovPlatform/protoflow/compare/0.4.2...0.4.3
+[0.4.2]: https://github.com/AsimovPlatform/protoflow/compare/0.4.1...0.4.2
+[0.4.1]: https://github.com/AsimovPlatform/protoflow/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/AsimovPlatform/protoflow/compare/0.3.1...0.4.0
 [0.3.1]: https://github.com/AsimovPlatform/protoflow/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/AsimovPlatform/protoflow/compare/0.2.2...0.3.0